pmh-13 - add admission services - add patient to AdmissionDto - add Admission entity

+++ src/main/java/ru/iteco/fmh/dao/repository/AdvertisementRepository.java
@@ -3,5 +3,5 @@ package ru.iteco.fmh.dao.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.iteco.fmh.model.Advertisement;
 
-public interface AdvertisementRepository extends JpaRepository<Advertisement, Long> {
+public interface AdvertisementRepository extends JpaRepository<Advertisement, Integer> {
 }
